requirements.txt
pymysql>=1.1.0
pyside6>=6.5.0
pyyaml>=6.0.1
cryptography>=41.0.1
typing-extensions>=4.7.1
nuitka>=1.8.0
apscheduler>=3.10.0

# 多数据库支持
psycopg2-binary>=2.9.0  # PostgreSQL 数据库驱动